server_function/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery==3.13.0
